requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.3
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
Flask==2.0.3
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.6.5
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
pip-review==1.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pyee==9.0.4
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.20.0
python3-openid==3.2.0
python3-saml==1.14.0
pytz==2022.1
PyYAML==6.0
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
slack==0.0.2
slack-sdk==3.15.2
slackclient==2.9.3
slackeventsapi==3.0.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
websocket-client==1.3.1
Werkzeug==2.0.3
whitenoise==5.3.0
xmlsec==1.3.12
yarl==1.7.2
zipp==3.7.0
